+++ apps/analytics/.eslintrc.json
