+++ apps/analytics/jest.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+  displayName: 'analytics',
+  preset: '../../jest.preset.js',
+  setupFiles: ['../../jest.setup.ts'],
+  globals: {
+    'ts-jest': {
+      tsconfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]s$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'js', 'html'],
+  coverageDirectory: '../../coverage/apps/analytics',
+};
