+++ apps/analytics/src/environments/environment.prod.ts
@@ -1,0 +1,3 @@
+export const environment = {
+  production: true,
+};
