+++ apps/analytics/tsconfig.app.json
