+++ apps/analytics/tsconfig.json
